Move API permission interfaces inward to db layer.
